refactor: move command system to base package
